+++ server/TWS Foundation/TWS Foundation.Quality/Q_Controllers/Business/Q_SituationsController.cs
@@ -1,14 +1,6 @@
-﻿using System.Net;
-
-using CSM_Foundation.Database.Entity.Depot.IDepot_View;
-
-using CSM_Security.Entities;
-
-using Microsoft.AspNetCore.Mvc.Testing;
+﻿using Microsoft.AspNetCore.Mvc.Testing;
 
 using TWS_Business.Entities;
-
-using TWS_Foundation.Middlewares.Frames;
 
 namespace TWS_Foundation.Quality.Q_Controllers.Business;
 
@@ -31,20 +23,6 @@ public class Q_SitutationsController
         };
     }
 
-    [Fact]
-    public async Task View() {
-        (HttpStatusCode Status, GenericFrame Response) = await Post("View", new ViewInput<TWS_Security.Sets.Account> {
-            Page = 1,
-            Range = 10,
-            Retroactive = false,
-        }, true);
 
-        Assert.Equal(HttpStatusCode.OK, Status);
-
-        View Estela = Framing<SuccessFrame<View>>(Response).Estela;
-        Assert.True(Estela.Records.Length > 0);
-        Assert.Equal(1, Estela.Page);
-        Assert.True(Estela.Pages > 0);
-    }
 
 }